tests/test_alpha_vantage_provider.py
"""Unit tests for the Alpha Vantage market data provider.

This test suite provides comprehensive coverage (93%) of the Alpha Vantage provider
implementation. It tests all major functionality including:

1. Historical Data:
   - Daily and intraday data retrieval
   - Date range filtering
   - Interval validation

2. Real-time Data:
   - Latest quote retrieval
   - Subscription management
   - Streaming simulation

3. Connection Management:
   - Connect/disconnect lifecycle
   - Session cleanup
   - Task management

4. Error Handling:
   - API errors
   - Invalid parameters
   - Connection state validation
   - Rate limiting
"""

import pytest
from datetime import datetime
from unittest.mock import patch, AsyncMock, MagicMock
import asyncio
from yarl import URL
from typing import Dict

import pandas as pd
from aioresponses import aioresponses
import aiohttp

from backend.services.market_data.alpha_vantage_provider import AlphaVantageProvider
from backend.services.market_data.base import MarketDataConfig, MarketDataCredentials

@pytest.mark.asyncio
class TestAlphaVantageProvider:
    @pytest.fixture
    def mock_session(self):
        """Create a mock session with proper async context manager support"""
        from unittest.mock import MagicMock
        
        mock = MagicMock()
        enter_mock = AsyncMock()
        enter_mock.json = AsyncMock()
        enter_mock.raise_for_status = AsyncMock()
        exit_mock = AsyncMock()
        
        mock.get.return_value.__aenter__ = enter_mock
        mock.get.return_value.__aexit__ = exit_mock
        mock.get.return_value.__aexit__.return_value = None
        
        return mock

    @pytest.fixture
    def provider(self, mock_session):
        config = MarketDataConfig(
            base_url="https://www.alphavantage.co/query",
            credentials=MarketDataCredentials(api_key="test_key"),
            websocket_url="not_used"
        )
        provider = AlphaVantageProvider(config)
        provider.session = mock_session
        return provider

    def _get_full_url(self, base_url: str, params: Dict[str, str]) -> str:
        """Helper to get full URL with params"""
        return str(URL(base_url).with_query(params))

    async def test_historical_data_daily(self, provider, mock_session):
        """Test daily historical data retrieval"""
        mock_response = {
            "Time Series (Daily)": {
                "2024-01-02": {
                    "1. open": "100.0",
                    "2. high": "101.0",
                    "3. low": "99.0",
                    "4. close": "100.5",
                    "5. volume": "1000000"
                },
                "2024-01-03": {
                    "1. open": "100.5",
                    "2. high": "102.0",
                    "3. low": "100.0",
                    "4. close": "101.5",
                    "5. volume": "1100000"
                }
            }
        }

        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        df = await provider.get_historical_data(
            symbol="AAPL",
            start_date=datetime(2024, 1, 1),
            interval="1d"
        )
        
        assert not df.empty
        assert len(df) == 2
        assert df.iloc[0]["close"] == 100.5
        assert df.iloc[1]["close"] == 101.5

        # Verify the request was made correctly
        mock_session.get.assert_called_once()
        args, kwargs = mock_session.get.call_args
        assert args[0] == "https://www.alphavantage.co/query"
        assert kwargs["params"]["function"] == "TIME_SERIES_DAILY"
        assert kwargs["params"]["symbol"] == "AAPL"

    async def test_historical_data_intraday(self, provider, mock_session):
        """Test intraday historical data retrieval"""
        mock_response = {
            "Time Series (5min)": {
                "2024-01-02 09:30:00": {
                    "1. open": "100.0",
                    "2. high": "100.2",
                    "3. low": "99.8",
                    "4. close": "100.1",
                    "5. volume": "10000"
                },
                "2024-01-02 09:35:00": {
                    "1. open": "100.1",
                    "2. high": "100.3",
                    "3. low": "100.0",
                    "4. close": "100.2",
                    "5. volume": "9500"
                }
            }
        }

        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        df = await provider.get_historical_data(
            symbol="AAPL",
            start_date=datetime(2024, 1, 2),
            interval="5min"
        )
        
        assert not df.empty
        assert len(df) == 2
        assert df.iloc[0]["close"] == 100.1
        assert df.iloc[1]["close"] == 100.2

        # Verify the request was made correctly
        mock_session.get.assert_called_once()
        args, kwargs = mock_session.get.call_args
        assert args[0] == "https://www.alphavantage.co/query"
        assert kwargs["params"]["function"] == "TIME_SERIES_INTRADAY"
        assert kwargs["params"]["symbol"] == "AAPL"
        assert kwargs["params"]["interval"] == "5min"

    async def test_latest_quote(self, provider, mock_session):
        """Test latest quote retrieval"""
        mock_response = {
            "Global Quote": {
                "01. symbol": "AAPL",
                "02. open": "100.0",
                "03. high": "101.0",
                "04. low": "99.0",
                "05. price": "100.5",
                "06. volume": "1000000",
                "07. latest trading day": "2024-01-02",
                "08. previous close": "99.5",
                "09. change": "1.0",
                "10. change percent": "1.0%"
            }
        }

        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        quote = await provider.get_latest_quote("AAPL")
        
        assert quote is not None
        assert quote["symbol"] == "AAPL"

        # Verify the request was made correctly
        mock_session.get.assert_called_once()
        args, kwargs = mock_session.get.call_args
        assert args[0] == "https://www.alphavantage.co/query"
        assert kwargs["params"]["function"] == "GLOBAL_QUOTE"
        assert kwargs["params"]["symbol"] == "AAPL"

    async def test_api_error(self, provider, mock_session):
        """Test API error handling"""
        mock_session.get.return_value.__aenter__.return_value.raise_for_status.side_effect = aiohttp.ClientResponseError(
            request_info=AsyncMock(),
            history=(),
            status=429
        )

        with pytest.raises(Exception):
            await provider.get_latest_quote("AAPL")

        # Verify the request was made
        mock_session.get.assert_called_once()

    async def test_invalid_interval(self, provider):
        """Test invalid interval handling"""
        with pytest.raises(ValueError, match="Unsupported interval: invalid"):
            await provider.get_historical_data(
                symbol="AAPL",
                start_date=datetime(2024, 1, 1),
                interval="invalid"
            )

    async def test_subscribe_unsubscribe(self, provider):
        """Test subscribe/unsubscribe functionality"""
        # Test subscribe
        symbols = ["AAPL", "MSFT"]
        await provider.subscribe(symbols)
        assert provider._subscribed_symbols == set(symbols)
        
        # Test unsubscribe
        await provider.unsubscribe(["AAPL"])
        assert provider._subscribed_symbols == {"MSFT"}
        
        # Test unsubscribe all
        await provider.unsubscribe(["MSFT"])
        assert not provider._subscribed_symbols

    async def test_stream_start_stop(self, provider, mock_session):
        """Test stream start and stop behavior"""
        mock_response = {
            "Global Quote": {
                "01. symbol": "AAPL",
                "05. price": "150.0",
                "06. volume": "1000000",
                "09. change": "2.0",
                "10. change percent": "1.5%"
            }
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response
        
        # Subscribe to a symbol
        await provider.subscribe(["AAPL"])
        
        # Start streaming
        provider._stop_streaming = False
        stream_task = asyncio.create_task(provider._stream_market_data())
        
        # Let it run for a bit
        await asyncio.sleep(0.1)
        
        # Stop streaming
        provider._stop_streaming = True
        await stream_task
        
        # Verify the request was made
        assert mock_session.get.called

    async def test_stream_error_handling(self, provider, mock_session):
        """Test error handling during streaming"""
        # Set up the mock to fail on first call then succeed on second
        mock_session.get.return_value.__aenter__.return_value.raise_for_status.side_effect = [
            aiohttp.ClientResponseError(
                request_info=AsyncMock(),
                history=(),
                status=429
            ),
            None
        ]
        
        mock_response = {
            "Global Quote": {
                "01. symbol": "AAPL",
                "05. price": "150.0",
                "06. volume": "1000000",
                "09. change": "2.0",
                "10. change percent": "1.5%"
            }
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response
        
        # Subscribe to a symbol
        await provider.subscribe(["AAPL"])
        
        # Start streaming
        provider._stop_streaming = False
        stream_task = asyncio.create_task(provider._stream_market_data())
        
        # Let it run for a bit
        await asyncio.sleep(0.1)
        
        # Stop streaming
        provider._stop_streaming = True
        await stream_task
        
        # Verify both requests were made
        assert mock_session.get.call_count >= 1

    async def test_connect_disconnect(self, provider):
        """Test connection management"""
        # Test connect
        await provider.connect()
        assert provider.session is not None
        assert not provider._stop_streaming
        assert provider._stream_task is None
        
        # Subscribe to trigger streaming
        await provider.subscribe(["AAPL"])
        assert provider._stream_task is not None
        assert not provider._stream_task.done()
        
        # Store task reference
        task = provider._stream_task
        
        # Test disconnect
        await provider.disconnect()
        assert provider.session is None
        assert provider._stop_streaming
        assert task.done()  # Check the stored task reference
        assert not provider._subscribed_symbols

    async def test_rate_limit_handling(self, provider, mock_session):
        """Test rate limit handling and backoff"""
        # Mock rate limit error response
        mock_session.get.return_value.__aenter__.return_value.raise_for_status.side_effect = [
            aiohttp.ClientResponseError(
                request_info=AsyncMock(),
                history=(),
                status=429
            ),
            None  # Second request succeeds
        ]
        
        mock_response = {
            "Global Quote": {
                "01. symbol": "AAPL",
                "02. open": "100.0",
                "03. high": "101.0",
                "04. low": "99.0",
                "05. price": "150.0",
                "06. volume": "1000000",
                "07. latest trading day": "2024-01-02",
                "08. previous close": "99.5",
                "09. change": "1.0",
                "10. change percent": "1.0%"
            }
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        # First request should fail with rate limit
        with pytest.raises(Exception):
            await provider.get_latest_quote("AAPL")

        # Second request should succeed after backoff
        quote = await provider.get_latest_quote("AAPL")
        assert quote is not None
        assert quote["symbol"] == "AAPL"
        assert quote["price"] == 150.0
        assert quote["volume"] == 1000000

    async def test_empty_response_handling(self, provider, mock_session):
        """Test handling of empty API responses"""
        mock_response = {
            "Time Series (Daily)": {}  # Empty data
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        df = await provider.get_historical_data(
            symbol="AAPL",
            start_date=datetime(2024, 1, 1),
            interval="1d"
        )
        
        assert df.empty
        assert isinstance(df, pd.DataFrame)

    async def test_malformed_data_handling(self, provider, mock_session):
        """Test handling of malformed API responses"""
        mock_response = {
            "Time Series (Daily)": {
                "2024-01-02": {
                    "1. open": "invalid",  # Invalid numeric data
                    "2. high": "101.0",
                    "3. low": "99.0",
                    "4. close": None,  # Missing data
                    "5. volume": "1000000"
                }
            }
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        with pytest.raises(ValueError):
            await provider.get_historical_data(
                symbol="AAPL",
                start_date=datetime(2024, 1, 1),
                interval="1d"
            )

    async def test_not_connected_error(self, provider):
        """Test error when making requests without connecting"""
        provider.session = None
        
        with pytest.raises(RuntimeError, match="Not connected to Alpha Vantage"):
            await provider.get_latest_quote("AAPL")
            
        with pytest.raises(RuntimeError, match="Not connected to Alpha Vantage"):
            await provider.get_historical_data(
                symbol="AAPL",
                start_date=datetime(2024, 1, 1)
            )

    async def test_invalid_api_key(self, provider, mock_session):
        """Test handling of invalid API key"""
        mock_session.get.return_value.__aenter__.return_value.raise_for_status.side_effect = aiohttp.ClientResponseError(
            request_info=AsyncMock(),
            history=(),
            status=401,
            message="Invalid API key"
        )

        with pytest.raises(Exception) as exc_info:
            await provider.get_latest_quote("AAPL")
        assert "401" in str(exc_info.value)

    async def test_missing_time_series_key(self, provider, mock_session):
        """Test handling of missing time series key in response"""
        mock_response = {
            "Error Message": "Invalid API call"
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        with pytest.raises(ValueError, match="API Error: Invalid API call"):
            await provider.get_historical_data(
                symbol="AAPL",
                start_date=datetime(2024, 1, 1),
                interval="1d"
            )

    async def test_intraday_interval_validation(self, provider):
        """Test validation of intraday intervals"""
        valid_intervals = ["1min", "5min", "15min", "30min", "1h"]
        invalid_intervals = ["2min", "10min", "2h"]
        
        # Test valid intervals
        for interval in valid_intervals:
            try:
                await provider.get_historical_data(
                    symbol="AAPL",
                    start_date=datetime(2024, 1, 1),
                    interval=interval
                )
            except Exception as e:
                assert not isinstance(e, ValueError), f"Valid interval {interval} raised ValueError"
        
        # Test invalid intervals
        for interval in invalid_intervals:
            with pytest.raises(ValueError, match=f"Unsupported interval: {interval}"):
                await provider.get_historical_data(
                    symbol="AAPL",
                    start_date=datetime(2024, 1, 1),
                    interval=interval
                )

    async def test_end_date_filtering(self, provider, mock_session):
        """Test end date filtering in historical data"""
        mock_response = {
            "Time Series (Daily)": {
                "2024-01-02": {
                    "1. open": "100.0",
                    "2. high": "101.0",
                    "3. low": "99.0",
                    "4. close": "100.5",
                    "5. volume": "1000000"
                },
                "2024-01-03": {
                    "1. open": "100.5",
                    "2. high": "102.0",
                    "3. low": "100.0",
                    "4. close": "101.5",
                    "5. volume": "1100000"
                },
                "2024-01-04": {
                    "1. open": "101.5",
                    "2. high": "103.0",
                    "3. low": "101.0",
                    "4. close": "102.5",
                    "5. volume": "1200000"
                }
            }
        }
        mock_session.get.return_value.__aenter__.return_value.json.return_value = mock_response

        # Test with end date
        df = await provider.get_historical_data(
            symbol="AAPL",
            start_date=datetime(2024, 1, 1),
            end_date=datetime(2024, 1, 3),
            interval="1d"
        )
        
        assert not df.empty
        assert len(df) == 2  # Should only include 01/02 and 01/03
        assert df.iloc[-1]["close"] == 101.5  # Last price should be from 01/03
